+++ indispensable-cosef/src/commonMain/kotlin/at/asitplus/signum/indispensable/cosef/CoseSigned.kt
@@ -2,22 +2,26 @@ package at.asitplus.signum.indispensable.cosef
 
 import at.asitplus.KmmResult
 import at.asitplus.catching
-import at.asitplus.signum.indispensable.*
-import at.asitplus.signum.indispensable.cosef.io.Base16Strict
+import at.asitplus.signum.indispensable.CryptoSignature
+import at.asitplus.signum.indispensable.contentEqualsIfArray
+import at.asitplus.signum.indispensable.contentHashCodeIfArray
+import at.asitplus.signum.indispensable.cosef.CoseSigned.Companion.create
 import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapper
 import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapperSerializer
 import at.asitplus.signum.indispensable.cosef.io.coseCompliantSerializer
-import at.asitplus.signum.indispensable.pki.X509Certificate
-import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
-import kotlinx.serialization.*
-import kotlinx.serialization.cbor.ByteString
-import kotlinx.serialization.cbor.CborArray
+import io.matthewnelson.encoding.base16.Base16
+import io.matthewnelson.encoding.core.Encoder.Companion.encodeToByteArray
+import kotlinx.serialization.ExperimentalSerializationApi
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.encodeToByteArray
 
 /**
  * Representation of a signed COSE_Sign1 object, i.e. consisting of protected header, unprotected header and payload.
  *
  * If the payload is a generic [ByteArray], then it will be serialized as-is. Should the payload be any other type,
- * the [CoseSignedSerializer] will tag it with 24 (see [RFC8949 3.4.5.1](https://www.rfc-editor.org/rfc/rfc8949.html#name-encoded-cbor-data-item)) during serialization.
+ * we will tag it with 24 (see
+ * [RFC8949 3.4.5.1](https://www.rfc-editor.org/rfc/rfc8949.html#name-encoded-cbor-data-item)) for serialization.
  * In order to prevent nested wrapping of the payload and the resulting type erasure
  * payloads of type [ByteStringWrapper] will be rejected.
  * In this case the payload could be handed over as the wrapped class itself or manually serialized to [ByteArray]
@@ -26,38 +30,17 @@ import kotlinx.serialization.cbor.CborArray
  */
 @OptIn(ExperimentalSerializationApi::class)
 @Serializable(with = CoseSignedSerializer::class)
-@CborArray
 @ConsistentCopyVisibility
 data class CoseSigned<P : Any?> internal constructor(
-    @ByteString
-    val protectedHeader: ByteStringWrapper<CoseHeader>,
-    val unprotectedHeader: CoseHeader?,
-    @ByteString
+    val protectedHeader: CoseHeader,
+    val unprotectedHeader: CoseHeader? = null,
     val payload: P?,
-    @ByteString
-    val rawSignature: ByteArray,
+    val signature: CryptoSignature.RawByteEncodable,
+    val wireFormat: CoseSignedBytes,
 ) {
-    @Throws(IllegalArgumentException::class)
-    constructor(
-        protectedHeader: CoseHeader,
-        unprotectedHeader: CoseHeader?,
-        payload: P?,
-        signature: CryptoSignature.RawByteEncodable,
-    ) : this(
-        protectedHeader = ByteStringWrapper(value = protectedHeader),
-        unprotectedHeader = unprotectedHeader,
-        payload = when (payload) {
-            is ByteStringWrapper<*> -> throw IllegalArgumentException("payload shall not be ByteStringWrapper")
-            else -> payload
-        },
-        rawSignature = signature.rawByteArray
-    )
 
-    val signature: CryptoSignature by lazy {
-        if (protectedHeader.value.usesEC() ?: unprotectedHeader?.usesEC() ?: (rawSignature.size < 2048))
-            CryptoSignature.EC.fromRawBytes(rawSignature)
-        else CryptoSignature.RSAorHMAC(rawSignature)
-    }
+    fun prepareCoseSignatureInput(externalAad: ByteArray = byteArrayOf()): ByteArray =
+        wireFormat.toCoseSignatureInput(externalAad)
 
     fun serialize(parameterSerializer: KSerializer<P>): ByteArray = coseCompliantSerializer
         .encodeToByteArray(CoseSignedSerializer(parameterSerializer), this)
@@ -74,22 +57,27 @@ data class CoseSigned<P : Any?> internal constructor(
             if (other.payload == null) return false
             if (!payload.contentEqualsIfArray(other.payload)) return false
         } else if (other.payload != null) return false
-        return rawSignature.contentEquals(other.rawSignature)
+        if (signature != other.signature) return false
+        if (wireFormat != other.wireFormat) return false
+
+        return true
     }
 
     override fun hashCode(): Int {
         var result = protectedHeader.hashCode()
         result = 31 * result + (unprotectedHeader?.hashCode() ?: 0)
         result = 31 * result + (payload?.contentHashCodeIfArray() ?: 0)
-        result = 31 * result + rawSignature.contentHashCode()
+        result = 31 * result + signature.hashCode()
+        result = 31 * result + wireFormat.hashCode()
         return result
     }
 
     override fun toString(): String {
-        return "CoseSigned(protectedHeader=${protectedHeader.value}," +
+        return "CoseSigned(protectedHeader=$protectedHeader," +
                 " unprotectedHeader=$unprotectedHeader," +
-                " payload=${if (payload is ByteArray) payload.encodeToString(Base16Strict) else payload}," +
-                " signature=${rawSignature.encodeToString(Base16Strict)})"
+                " payload=${if (payload is ByteArray) payload.encodeToByteArray(Base16()) else payload}," +
+                " signature=$signature," +
+                " wireFormat=$wireFormat)"
     }
 
     companion object {
@@ -99,85 +87,63 @@ data class CoseSigned<P : Any?> internal constructor(
             }
 
         /**
-         * Called by COSE signing implementations to get the bytes that will be
-         * used as the input for signature calculation of a `COSE_Sign1` object
+         * Use this method to create a new [CoseSigned] object with correct [CoseSigned.wireFormat] set.
          */
-        fun <P : Any> prepareCoseSignatureInput(
+        @Throws(IllegalArgumentException::class)
+        fun <P : Any> create(
             protectedHeader: CoseHeader,
+            unprotectedHeader: CoseHeader? = null,
             payload: P?,
-            serializer: KSerializer<P>,
+            signature: CryptoSignature.RawByteEncodable,
+            payloadSerializer: KSerializer<P>,
+        ): CoseSigned<P> = CoseSigned<P>(
+            protectedHeader = protectedHeader,
+            unprotectedHeader = unprotectedHeader,
+            payload = payload,
+            signature = signature,
+            wireFormat = CoseSignedBytes(
+                protectedHeader = coseCompliantSerializer.encodeToByteArray(protectedHeader),
+                unprotectedHeader = unprotectedHeader,
+                payload = payload.toRawPayload(payloadSerializer),
+                rawSignature = signature.rawByteArray
+            ),
+        )
+
+        /**
+         * Use this method to prepare a [CoseSignatureInput] object to calculate the signature,
+         * and then call [create] to create a [CoseSigned] object.
+         */
+        @Throws(IllegalArgumentException::class)
+        fun <P : Any> prepare(
+            protectedHeader: CoseHeader,
             externalAad: ByteArray = byteArrayOf(),
-        ): ByteArray = CoseSignatureInput(
+            payload: P?,
+            payloadSerializer: KSerializer<P>,
+        ): CoseSignatureInput = CoseSignatureInput(
             contextString = "Signature1",
-            protectedHeader = ByteStringWrapper(protectedHeader),
+            protectedHeader = coseCompliantSerializer.encodeToByteArray<CoseHeader>(protectedHeader),
             externalAad = externalAad,
-            payload = when (payload) {
-                null -> null
-                is ByteArray -> payload
-                else -> coseCompliantSerializer.encodeToByteArray(
-                    ByteStringWrapperSerializer(serializer),
-                    ByteStringWrapper(payload)
-                )
-            },
-        ).serialize()
-    }
-}
+            payload = payload.toRawPayload(payloadSerializer),
+        )
 
-fun CoseHeader.usesEC(): Boolean? = algorithm?.algorithm?.let { it is SignatureAlgorithm.ECDSA }
-    ?: certificateChain?.let { X509Certificate.decodeFromDerOrNull(it)?.publicKey is CryptoPublicKey.EC }
+        /**
+         * If [this] is a [ByteArray], use it as is, otherwise encode it as a [ByteStringWrapper], with CBOR tag 24
+         */
+        private fun <P : Any> P?.toRawPayload(payloadSerializer: KSerializer<P>): ByteArray = when (this) {
+            is ByteArray -> this
+            is Nothing -> byteArrayOf()
+            is ByteStringWrapper<*> -> throw IllegalArgumentException("Payload must not be a ByteStringWrapper")
+            is P -> coseCompliantSerializer.encodeToByteArray<ByteStringWrapper<P>>(
+                ByteStringWrapperSerializer(payloadSerializer),
+                ByteStringWrapper(this)
+            ).wrapInCborTag(24)
 
-
-@OptIn(ExperimentalSerializationApi::class)
-@Serializable
-@CborArray
-data class CoseSignatureInput(
-    val contextString: String,
-    @ByteString
-    val protectedHeader: ByteStringWrapper<CoseHeader>,
-    @ByteString
-    val externalAad: ByteArray,
-    @ByteString
-    val payload: ByteArray?,
-) {
-    fun serialize() = coseCompliantSerializer.encodeToByteArray(this)
-
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other == null || this::class != other::class) return false
-
-        other as CoseSignatureInput
-
-        if (contextString != other.contextString) return false
-        if (protectedHeader != other.protectedHeader) return false
-        if (!externalAad.contentEquals(other.externalAad)) return false
-        if (payload != null) {
-            if (other.payload == null) return false
-            if (!payload.contentEquals(other.payload)) return false
-        } else if (other.payload != null) return false
-
-        return true
-    }
-
-    override fun hashCode(): Int {
-        var result = contextString.hashCode()
-        result = 31 * result + protectedHeader.hashCode()
-        result = 31 * result + externalAad.contentHashCode()
-        result = 31 * result + (payload?.contentHashCode() ?: 0)
-        return result
-    }
-
-    override fun toString(): String {
-        return "CoseSignatureInput(contextString='$contextString'," +
-                " protectedHeader=${protectedHeader.value}," +
-                " externalAad=${externalAad.encodeToString(Base16Strict)}," +
-                " payload=${payload?.encodeToString(Base16Strict)})"
-    }
-
-
-    companion object {
-        fun deserialize(it: ByteArray) = catching {
-            coseCompliantSerializer.decodeFromByteArray<CoseSignatureInput>(it)
+            else -> byteArrayOf()
         }
+
+        private fun ByteArray.wrapInCborTag(tag: Byte) = byteArrayOf(0xd8.toByte()) + byteArrayOf(tag) + this
+
     }
 }
+